Move providers into their own component
